backend/requirements.txt
# Flask Framework
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# MongoDB
pymongo==4.6.0
dnspython==2.4.2

# Gemini AI
google-generativeai==0.3.2

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.1.0
python-dotenv==1.0.0

# File Processing
Pillow==10.1.0

# Utilities
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# Development
black==23.12.0
flake8==6.1.0
